Fix idle animation: normalisera IdleCounter, slå ner flapp-hastighet, lägg till lyft
Tre buggar åtgärdade:

1. IdleCounter++ → IdleCounter += elapsed * 60f
   Räknaren var frames-baserad och träffade IdleTimeout=200 på ~1.6s vid 120fps
   (original: ~3.3s vid 60fps). Normalisering via 60fps-ekvivalenta enheter,
   samma mönster som DynamicItem._subTick-fixet.
   GameContext.IdleCounter ändrad från int till float.

2. Idle-flapp: GraphicCounter % 2 → Timer < 0.2f
   GraphicCounter % 2 bytte ram var 0.1s = 10 Hz (glitchigt).
   Timer-baserad halvdelning ger 0.4s-cykel halvt/halvt = 2.5 Hz.

3. Lyft: kontinuerlig vy -= 20.1f*elapsed → enstaka impuls vy = -2.0f
   Ursprungliga versionen tog ut sig mot GravityNormal (20.0f) i samma
   frame — netto ~noll rörelse. Nu appliceras ett enda vertikalt impuls
   i lyftzonen (counter 220-221) och gravitation tar vid naturligt.

+++ FrostyPlatformer/Core/GameContext.cs
@@ -96,8 +96,8 @@ namespace FrostyPlatformer.Core
         /// <summary>Speltid vid avslutsskärmen — fryst när spelaren når slutet.</summary>
         public TimeSpan EndTotalTime { get; set; }
 
-        /// <summary>Räknare för inaktivitet i idle-detektion.</summary>
-        public int IdleCounter { get; set; }
+        /// <summary>Räknare för inaktivitet i idle-detektion (normaliserat till 60 fps-ekvivalenta enheter).</summary>
+        public float IdleCounter { get; set; }
 
         // ─────────────────────────────────────────────
         // Navigationsläge för menyer (delas mellan MenuState, SettingsState, PauseState)
